# Conflicts:
#	onnxruntime/python/tools/quantization/registry.py
#	tools/ci_build/github/azure-pipelines/templates/c-api-linux-cpu.yml

+++ orttraining/orttraining/test/python/orttraining_test_ortmodule_api.py
@@ -39,7 +39,7 @@ from onnxruntime.training.ortmodule import (
 )
 from onnxruntime.training.ortmodule._custom_gradient_registry import register_gradient
 
-DEFAULT_OPSET = 14
+DEFAULT_OPSET = 15
 
 
 # PyTorch model definitions for tests
@@ -1326,51 +1326,56 @@ def test_gradient_correctness_reducesum(dim, keepdim):
 # Before PyTorch 1.11.0, the exporter will fail to register symbolic with non-empty domain.
 @pytest.mark.skipif(Version(torch.__version__) < Version("1.11.0"), reason="PyTorch 1.10 incompatible")
 @pytest.mark.parametrize("dim", [0, 1, -1])
-def test_gradient_correctness_chunk(dim):
+@pytest.mark.parametrize("chunks", [1, 3])
+def test_gradient_correctness_chunk(dim, chunks):
     class NeuralNetChunk(torch.nn.Module):
         def __init__(self, dim):
             super(NeuralNetChunk, self).__init__()
             self.dim = dim
 
         def forward(self, input):
-            return input.chunk(3, dim=self.dim)
+            return input.chunk(chunks, dim=self.dim)
 
     device = "cuda"
     pt_model = NeuralNetChunk(dim).to(device)
-    ort_model = ORTModule(copy.deepcopy(pt_model), DebugOptions(save_onnx=True, onnx_prefix="chunk_model"))
+    ort_model = ORTModule(copy.deepcopy(pt_model), DebugOptions(save_onnx=(chunks > 1), onnx_prefix="chunk_model"))
 
     def run_step(model, input):
-        y1, y2, y3 = model(input)
-        loss = y1.sum() + y2.sum() + y3.sum()
+        results = model(input)
+        loss = results[0].sum()
+        for i in range(1, len(results)):
+            loss = loss + results[i].sum()
         loss.backward()
-        return y1, y2, y3
+        return results
 
     N, D, H = 16, 17, 18
     for _ in range(10):
-        input = torch.rand((N, D, H), device=device, requires_grad=True)
-        pt_y1, pt_y2, pt_y3 = run_step(pt_model, input)
-        ort_y1, ort_y2, ort_y3 = run_step(ort_model, input)
+        pt_input = torch.rand((N, D, H), device=device, requires_grad=True)
+        ort_input = copy.deepcopy(pt_input)
+        pt_results = run_step(pt_model, pt_input)
+        ort_results = run_step(ort_model, ort_input)
 
-        _test_helpers.assert_values_are_close(ort_y1, pt_y1)
-        _test_helpers.assert_values_are_close(ort_y2, pt_y2)
-        _test_helpers.assert_values_are_close(ort_y3, pt_y3)
-        _test_helpers.assert_gradients_match_and_reset_gradient(ort_model, pt_model)
+        assert len(ort_results) == len(pt_results)
+        for i in range(len(ort_results)):
+            _test_helpers.assert_values_are_close(ort_results[i], pt_results[i])
+        _test_helpers.assert_values_are_close(ort_input.grad, pt_input.grad)
 
-    assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_torch_exported_training.onnx"))
-    assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_optimized_training.onnx"))
-    assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_optimized_pre_grad_training.onnx"))
-    assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_execution_model_training.onnx"))
-    model = onnx.load(os.path.join(os.getcwd(), "chunk_model_torch_exported_training.onnx"))
-    has_split = False
-    for node in model.graph.node:
-        if node.op_type == "Split":
-            has_split = True
-            break
-    assert has_split
-    os.remove(os.path.join(os.getcwd(), "chunk_model_torch_exported_training.onnx"))
-    os.remove(os.path.join(os.getcwd(), "chunk_model_optimized_training.onnx"))
-    os.remove(os.path.join(os.getcwd(), "chunk_model_optimized_pre_grad_training.onnx"))
-    os.remove(os.path.join(os.getcwd(), "chunk_model_execution_model_training.onnx"))
+    if chunks > 1:
+        assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_torch_exported_training.onnx"))
+        assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_optimized_training.onnx"))
+        assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_optimized_pre_grad_training.onnx"))
+        assert os.path.exists(os.path.join(os.getcwd(), "chunk_model_execution_model_training.onnx"))
+        model = onnx.load(os.path.join(os.getcwd(), "chunk_model_torch_exported_training.onnx"))
+        has_split = False
+        for node in model.graph.node:
+            if node.op_type == "Split":
+                has_split = True
+                break
+        assert has_split
+        os.remove(os.path.join(os.getcwd(), "chunk_model_torch_exported_training.onnx"))
+        os.remove(os.path.join(os.getcwd(), "chunk_model_optimized_training.onnx"))
+        os.remove(os.path.join(os.getcwd(), "chunk_model_optimized_pre_grad_training.onnx"))
+        os.remove(os.path.join(os.getcwd(), "chunk_model_execution_model_training.onnx"))
 
 
 # In PyTorch 1.11.0, there is issue during reduce node shape handling for exporter, so any sub-graph that
@@ -1623,39 +1628,6 @@ def test_numpy_T(input_shape):
     ort_prediction = run_step(ort_model, ort_input)
 
     _test_helpers.assert_values_are_close(ort_prediction, pt_prediction)
-
-
-def test_gradient_correctness_bce_with_logits():
-    class NeuralNetBCEWithLogitsLoss(torch.nn.Module):
-        def __init__(self, input_size, hidden_size):
-            super(NeuralNetBCEWithLogitsLoss, self).__init__()
-            self.linear = torch.nn.Linear(input_size, hidden_size)
-
-        def forward(self, input, target):
-            loss_fct = torch.nn.BCEWithLogitsLoss()
-            return loss_fct(self.linear(input), target)
-
-    N, D, H = 16, 256, 128
-    device = "cuda"
-    pt_model = NeuralNetBCEWithLogitsLoss(D, H).to(device)
-    ort_model = ORTModule(copy.deepcopy(pt_model))
-
-    def run_step(model, input, target):
-        prediction = model(input, target)
-        loss = prediction.sum()
-        loss.backward()
-        return prediction
-
-    for _ in range(10):
-        pt_input = torch.rand((N, D), device=device, requires_grad=True)
-        ort_input = copy.deepcopy(pt_input)
-        pt_target = torch.rand((N, H), device=device)
-        ort_target = copy.deepcopy(pt_target)
-        pt_prediction = run_step(pt_model, pt_input, pt_target)
-        ort_prediction = run_step(ort_model, ort_input, ort_target)
-
-        _test_helpers.assert_values_are_close(ort_prediction, pt_prediction)
-        _test_helpers.assert_values_are_close(ort_input.grad, pt_input.grad)
 
 
 def test_gradient_correctness_cast_chain():
@@ -5107,7 +5079,7 @@ def test_sigmoid_grad_opset13():
     old_opst_cst = ortmodule.ONNX_OPSET_VERSION
     old_opset = os.getenv("ORTMODULE_ONNX_OPSET_VERSION", None)
     os.environ["ORTMODULE_ONNX_OPSET_VERSION"] = "13"
-    assert ortmodule.ONNX_OPSET_VERSION == 14
+    assert ortmodule.ONNX_OPSET_VERSION == 15
 
     ort_model = ORTModule(copy.deepcopy(pt_model))
 
@@ -5137,7 +5109,7 @@ def test_sigmoid_grad_opset13():
     ortmodule.ONNX_OPSET_VERSION = old_opst_cst
 
 
-@pytest.mark.parametrize("opset_version", [12, 13, 14])
+@pytest.mark.parametrize("opset_version", [12, 13, 14, 15])
 def test_opset_version_change(opset_version):
     device = "cuda"
